+++ modules/activiti-engine/src/test/java/org/activiti/examples/bpmn/tasklistener/MyTransactionalOperationTransactionDependentTaskListener.java
@@ -12,13 +12,14 @@
  */
 package org.activiti.examples.bpmn.tasklistener;
 
+import java.util.List;
+import java.util.Map;
+
 import org.activiti.bpmn.model.Task;
 import org.activiti.engine.HistoryService;
 import org.activiti.engine.history.HistoricProcessInstance;
 import org.activiti.engine.impl.context.Context;
-
-import java.util.List;
-import java.util.Map;
+import org.activiti.engine.impl.history.HistoryLevel;
 
 /**
  * @author Yvo Swillens
@@ -29,10 +30,12 @@ public class MyTransactionalOperationTransactionDependentTaskListener extends Cu
   public void notify(String processInstanceId, String executionId, Task task, Map<String, Object> executionVariables, Map<String, Object> customPropertiesMap) {
     super.notify(processInstanceId, executionId, task, executionVariables, customPropertiesMap);
 
-    HistoryService historyService = Context.getCommandContext().getProcessEngineConfiguration().getHistoryService();
-
-    // delete first historic instance
-    List<HistoricProcessInstance> historicProcessInstances = historyService.createHistoricProcessInstanceQuery().list();
-    historyService.deleteHistoricProcessInstance(historicProcessInstances.get(0).getId());
+    if (Context.getCommandContext().getProcessEngineConfiguration().getHistoryLevel().isAtLeast(HistoryLevel.ACTIVITY)) {
+      HistoryService historyService = Context.getCommandContext().getProcessEngineConfiguration().getHistoryService();
+  
+      // delete first historic instance
+      List<HistoricProcessInstance> historicProcessInstances = historyService.createHistoricProcessInstanceQuery().list();
+      historyService.deleteHistoricProcessInstance(historicProcessInstances.get(0).getId());
+    }
   }
 }
